+++ src/scitacean/_html_repr/_dataset_html.py
@@ -156,10 +156,8 @@ def _get_fields(dset: Dataset) -> List[Field]:
 
 
 def _check_error(field: Dataset.Field, validation: Dict[str, str]) -> Optional[str]:
-    if field.name in validation:
-        # TODO validation uses model names (camelCase)
-        return validation[field.name]
-    return None
+    field_spec = next(filter(lambda f: f.name == field.name, Dataset.fields()))
+    return validation.get(field_spec.scicat_name, None)
 
 
 def _validate(dset: Dataset) -> Dict[str, str]:
@@ -209,6 +207,9 @@ def _human_readable_size(size_in_bytes: int) -> str:
 def _row_highlight_classes(field: Field) -> str:
     if field.required and field.value is None:
         return "cean-missing-value"
-    if field.error:
+    # Do not flag read-only fields with a value as errors.
+    # Validation is geared towards uploading where such fields must be None.
+    # But here, we don't want to flag downloaded datasets as bad because of this.
+    if field.error and not (field.read_only and field.error.startswith("Extra inputs")):
         return "cean-error"
     return ""